rpcclient: Remove dcrwallet methods
Methods calls on the Client type for performing dcrwallet JSON-RPCs
are being removed in order to simplify the dependency graphs between
projects. These calls are moving to a package in the dcrwallet module
that will abstract the actual connection and calling logic through an
interface (which rpcclient.Client will satisfy).

All callbacks for dcrwallet notifications have been removed.
Dcrwallet has not produced JSON-RPC notifications for several years at
this point.

The example program connecting to dcrwallet is also removed here. If
the dependencies of this program appeared in the go.mod (they don't
currently, as go-spew is missing), then rpcclient would be right back
to requiring the dcrwallet module. This example program may reappear
in the new wallet package. Any pointers towards the new wallet
package should be done in comments, not code that will update the
modules.

+++ rpcserver_test.go
@@ -85,7 +85,7 @@ func testGetBlockHash(ctx context.Context, r *rpctest.Harness, t *testing.T) {
 		t.Fatalf("call to getinfo failed: %v", err)
 	}
 
-	blockHash, err := r.Node.GetBlockHash(ctx, int64(info.Blocks))
+	blockHash, err := r.Node.GetBlockHash(ctx, info.Blocks)
 	if err != nil {
 		t.Fatalf("Call to `getblockhash` failed: %v", err)
 	}
